Factor out common test tasks
